Refactor OngoingCall module for greater speed
Instead of hiding all the ETS operations inside a GenServer, which can
be a bottleneck, this commit runs the lookup operation in the client
process. Save and delete are still inside the GenServer to ensure they
are done sequentially, however, they could be moved outside also.

To accomplish this, the ETS table has to have a name. Also, ExActor
seemed unnecessary for the simple GenServer here, so I removed it also.