Only sign objects present in hex package
I want you to be able to cd to the deps/telephonist folder in your
project and run `keybase dir verify`, and get a successful result. This
commit doesn’t bother signing the files that Hex ignores, so that the
verification will succeed. Otherwise, it would fail because of the
missing files.