Process calls in current process
Since `CallProcessor` just `awaits` the result of the `Task` anyway, the
current implementation is actually _slower_. This is because Elixir has
to spin up a new process to handle the call, and then receive the
output. It's much simpler to just handle the call in the current
process, and avoid the overhead of starting a pointless new process.